Tweaked Api response model field types. All integer types are now uint or ulong.

+++ SteamWebRequest/SteamApiClient/Models/Dota/Cosmetics/EquipedItem.cs
@@ -3,6 +3,6 @@
     public class EquipedItem
     {
         public ulong DefIndex { get; set; }
-        public byte style { get; set; }
+        public uint style { get; set; }
     }
 }
