+++ SteamWebRequest/SteamApiClient/Models/Dota/Match/LiveLeagueMatchScoreboard.cs
@@ -7,7 +7,7 @@ namespace SteamApi.Models.Dota
         public double Duration { get; set; }
 
         [JsonProperty("roshan_respawn_timer")]
-        public ushort RoshanRespawnTimer { get; set; }
+        public uint RoshanRespawnTimer { get; set; }
 
         public LiveLeagueMatchScoreboardTeam Dire { get; set; }
         public LiveLeagueMatchScoreboardTeam Radiant { get; set; }
