+++ SteamWebRequest/SteamApiClient/Models/Dota/Match/MatchHistoryResponse.cs
@@ -6,10 +6,10 @@ namespace SteamApi.Models.Dota
     public class MatchHistoryResponse
     {
         [JsonProperty("total_results")]
-        public ushort TotalCount { get; set; }
+        public uint TotalCount { get; set; }
 
         [JsonProperty("results_remaining")]
-        public ushort Remaining { get; set; }
+        public uint Remaining { get; set; }
 
         public byte Status { get; set; }
         public string StatusDetail { get; set; }
