+++ SteamWebRequest/SteamApiClient/Models/Dota/ProDota/Tournament/TournamentTeam.cs
@@ -4,12 +4,12 @@ namespace SteamApi.Models.Dota
 {
     public class TournamentTeam
     {
-        public int TeamId { get; set; }
+        public uint TeamId { get; set; }
         public string Name { get; set; }
         public double Payout { get; set; }
 
         [JsonProperty("games_won")]
-        public byte GamesWon { get; set; }
+        public uint GamesWon { get; set; }
 
         [JsonProperty("url_logo")]
         public string UrlLogo { get; set; }
